requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
numpy
flask-cors
matplotlib
